message-hub/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
requests==2.32.3
kafka-python==2.0.2
pytest==8.0.0
